+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Preserve existing Content-Encoding when applying trailer checksum.
+
 3.241.1 (2026-01-06)
 ------------------